requirements.txt
amqp==5.2.0
asgi-tools==0.76.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
biopython==1.83
celery==5.3.6
cffi==1.16.0
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contourpy==1.2.1
cron-descriptor==1.4.3
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.2
Django==4.2.11
django-bootstrap-v5==1.0.11
django-celery-beat==2.6.0
django-rename-app==0.1.7
django-timezone-field==6.1.0
et-xmlfile==1.1.0
fonttools==4.51.0
http-router==4.1.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kaleido==0.2.1
kiwisolver==1.4.5
kombu==5.3.7
matplotlib==3.8.4
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.20.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
scipy==1.13.0
service-identity==24.1.0
setuptools==69.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
tenacity==8.2.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
zope.interface==6.3
